move copy functions to scan
